LayoutTests/inspector/timeline-parse-html.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="timeline-test.js"></script>
<script>

function analyzeTimelineData(timelineRecords) 
{
    // Uncomment to debugging the list of data returned.
    // dumpTimelineRecords(timelineRecords);

    // Search for the first ParseHTML record you can find
    var numRecords = timelineRecords.length;
    for (var i = 0 ; i < numRecords; ++i) {
        var record = timelineRecords[i];
        if (record.type === timelineAgentRecordType.ParseHTML) {
            printTimelineRecordProperties(record);
            break;
        }
    }
}

function doit() 
{
    retrieveTimelineData(analyzeTimelineData);
}

</script>
</head>

<body onload="onload()">
<p>
Tests the Timeline API instrumentation of ParseHTML 
</p>

</body>
</html>
