LayoutTests/inspector/timeline-parse-html-expected.txt

Tests the Timeline API instrumentation of ParseHTML

ParseHTML Properties:
+ startTime : * DEFINED *
+ children : * DEFINED *
+ endTime : * DEFINED *
+ type : 4
